+++ lib/src/enums/query_type.dart
@@ -7,4 +7,6 @@ enum QueryType {
   spaces,
   counters,
   reports,
+  attachments,
+  attachments_aggregation,
 }