+++ lib/src/models/request/action_request.dart
@@ -1,9 +1,13 @@
 import 'package:dmart/src/enums/request_type.dart';
 import 'package:dmart/src/enums/resource_type.dart';
 
+/// ActionRequest is a class that represents a request to the API to perform an action on a resource.
 class ActionRequest {
+  /// The name of the space in which the resource resides.
   final String spaceName;
+  /// The type of request to perform.
   final RequestType requestType;
+  /// The records to perform the action on.
   final List<ActionRequestRecord> records;
 
   ActionRequest({
@@ -22,11 +26,17 @@ class ActionRequest {
   }
 }
 
+/// ActionRequestRecord is a class that represents a record in an ActionRequest.
 class ActionRequestRecord {
+  /// The type of resource to perform the action on.
   final ResourceType resourceType;
+  /// The shortname of the resource.
   String shortname;
+  /// The subpath of the resource.
   final String subpath;
+  /// The attributes to perform the action with.
   final Map<String, dynamic> attributes;
+  /// The attachments to perform the action with.
   final Map<ResourceType, List<dynamic>>? attachments;
 
   ActionRequestRecord({