+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+## 1.0.9
+- Implementing dmart features up to dmart@1.3.5+
+- Extending ResourceType: +reaction
+- Extending QueryType: +update, +delete
+- Extending dmart init to allow dio configuration.
+- remove deprecated feature `branch`.
+
 ## 1.0.7
 
-- Implementing dmart features up to dmart@1.1.12.
+- Implementing dmart features up to dmart@1.1.12
 
 ## 1.0.0